changed a few things on the hardware list
changed hardware the complete size to 355x355x350, as most beds sold in the market are 355x355mm instead of 350x350mm. so i prefer to just add 5mm overrall, also i made a mistake and calculate just 14 bars of 585, and it was actually 16 bars, so i changed on the list to 16 bars of 590mm, also changed the sized on 2020 gantry to accomodate this changes.